Fix schema dictionary used during robot status override
